requirements.txt
torch==2.7.0
torchvision==0.22.0
pillow==11.2.1
opencv-python==4.11.0.86
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
requests==2.32.3
